pyproject.toml
[tool.poetry]
name = "async-image-processing"
version = "0.1.0"
description = "Async Image Processing"
authors = ["Ali Naseri <alinaseri.ut@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.14"
fastapi = {extras = ["standard"], version = "^0.122.0"}
sqlalchemy = "^2.0.44"
python-magic = "^0.4.27"
psycopg2-binary = "^2.9.11"
boto3 = "^1.41.5"
pydantic-settings = "^2.12.0"
isort = "^7.0.0"
black = "^25.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
